updated nginx version and added luarocks

+++ Dockerfile
@@ -2,13 +2,12 @@ FROM alpine:3.5
 
 MAINTAINER Madhukar Thota "madhukar.thota@gmail.com"
 
-ENV NGINX_VERSION 1.11.10
+ENV NGINX_VERSION 1.12.0
 ENV DEVEL_KIT_MODULE_VERSION 0.3.0
-ENV LUA_MODULE_VERSION 0.10.8
+ENV LUAJIT_VERSION 2.1.0-beta2
 ENV LUAROCKS_VERSION 2.4.2
-
-ENV LUAJIT_LIB=/usr/lib
-ENV LUAJIT_INC=/usr/include/luajit-2.1
+ENV LUAJIT_LIB /usr/local/lib
+ENV LUAJIT_INC /usr/local/include/luajit-2.1
 
 RUN GPG_KEYS=B0F4253373F8F6F510D42178520A9993A1C052F8 \
 	&& CONFIG="\
@@ -42,26 +41,28 @@ RUN GPG_KEYS=B0F4253373F8F6F510D42178520A9993A1C052F8 \
 		--with-http_auth_request_module \
 		--with-http_xslt_module=dynamic \
 		--with-http_image_filter_module=dynamic \
-		--with-http_geoip_module \
+		--with-http_geoip_module=dynamic \
 		--with-http_perl_module=dynamic \
 		--with-threads \
 		--with-stream \
 		--with-stream_ssl_module \
+		--with-stream_ssl_preread_module \
+		--with-stream_realip_module \
 		--with-stream_geoip_module=dynamic \
 		--with-http_slice_module \
 		--with-mail \
 		--with-mail_ssl_module \
+		--with-compat \
 		--with-file-aio \
 		--with-http_v2_module \
-		--with-ipv6 \
+		--with-ld-opt="-Wl,--strip-debug,-rpath,${LUAJIT_LIB}" \
 		--add-module=/usr/src/ngx_devel_kit-$DEVEL_KIT_MODULE_VERSION \
-		--add-module=/usr/src/lua-nginx-module-$LUA_MODULE_VERSION \
+		--add-module=/usr/src/lua-nginx-module \
 	" \
 	&& addgroup -S nginx \
 	&& adduser -D -S -h /var/cache/nginx -s /sbin/nologin -G nginx nginx \
-	&& apk update \ 
+    && apk update \
 	&& apk add --no-cache --virtual .build-deps \
-	        git \
 		gcc \
 		libc-dev \
 		make \
@@ -75,28 +76,64 @@ RUN GPG_KEYS=B0F4253373F8F6F510D42178520A9993A1C052F8 \
 		gd-dev \
 		geoip-dev \
 		perl-dev \
-		luajit-dev \
-	&& curl -fSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz -o nginx.tar.gz \
-	&& curl -fSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz.asc  -o nginx.tar.gz.asc \
-	&& curl -fSL https://github.com/simpl/ngx_devel_kit/archive/v$DEVEL_KIT_MODULE_VERSION.tar.gz -o ndk.tar.gz \
-	&& curl -fSL https://github.com/openresty/lua-nginx-module/archive/v$LUA_MODULE_VERSION.tar.gz -o lua.tar.gz \
-	&& curl -fSL https://github.com/luarocks/luarocks/archive/v${LUAROCKS_VERSION}.tar.gz -o luarocks.tar.gz \
+	&& apk add --no-cache \
+	    git \
+		bash \
+		openssl \
+        build-base \
+        curl \
+        gd \
+        geoip \
+        libgcc \
+        libxslt \
+        linux-headers \
+        make \
+        perl \
+        unzip \
+        zlib \
+	&& curl -fsSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz -o nginx.tar.gz \
+	&& curl -fsSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz.asc  -o nginx.tar.gz.asc \
+	&& curl -fsSL https://github.com/simpl/ngx_devel_kit/archive/v$DEVEL_KIT_MODULE_VERSION.tar.gz -o ndk.tar.gz \
+	&& git clone https://github.com/openresty/lua-nginx-module.git /usr/src/lua-nginx-module \
+	&& curl -fsSL http://luajit.org/download/LuaJIT-${LUAJIT_VERSION}.tar.gz -o luajit.tar.gz \
+	&& curl -fsSL https://github.com/luarocks/luarocks/archive/v${LUAROCKS_VERSION}.tar.gz -o luarocks.tar.gz \
 	&& export GNUPGHOME="$(mktemp -d)" \
-	&& gpg --keyserver ha.pool.sks-keyservers.net --recv-keys "$GPG_KEYS" \
-	&& gpg --batch --verify nginx.tar.gz.asc nginx.tar.gz \
+	&& found=''; \
+	for server in \
+		ha.pool.sks-keyservers.net \
+		hkp://keyserver.ubuntu.com:80 \
+		hkp://p80.pool.sks-keyservers.net:80 \
+		pgp.mit.edu \
+	; do \
+		echo "Fetching GPG key $GPG_KEYS from $server"; \
+		gpg --keyserver "$server" --keyserver-options timeout=10 --recv-keys "$GPG_KEYS" && found=yes && break; \
+	done; \
+	test -z "$found" && echo >&2 "error: failed to fetch GPG key $GPG_KEYS" && exit 1; \
+	gpg --batch --verify nginx.tar.gz.asc nginx.tar.gz \
 	&& rm -r "$GNUPGHOME" nginx.tar.gz.asc \
 	&& mkdir -p /usr/src \
 	&& tar -zxC /usr/src -f nginx.tar.gz \
 	&& tar -zxC /usr/src -f ndk.tar.gz \
-	&& tar -zxC /usr/src -f lua.tar.gz \
+	&& tar -zxC /tmp -f luajit.tar.gz \
 	&& tar -zxC /tmp -f luarocks.tar.gz \
-	&& rm nginx.tar.gz ndk.tar.gz lua.tar.gz luarocks.tar.gz \
+	&& rm nginx.tar.gz ndk.tar.gz luajit.tar.gz luarocks.tar.gz \
+	&& cd /tmp/LuaJIT-${LUAJIT_VERSION} \
+	&& make -j$(getconf _NPROCESSORS_ONLN) \
+	&& make install \
+	&& ln -sf /usr/local/bin/luajit-* /usr/local/bin/lua \
+	&& cd /tmp/luarocks-* \
+	&& ./configure \
+		--with-lua-include=${LUAJIT_INC} \
+		--lua-suffix=jit-2.1.0-beta2 \
+	&& make -j$(getconf _NPROCESSORS_ONLN) \
+	&& make install \
 	&& cd /usr/src/nginx-$NGINX_VERSION \
 	&& ./configure $CONFIG --with-debug \
 	&& make -j$(getconf _NPROCESSORS_ONLN) \
 	&& mv objs/nginx objs/nginx-debug \
 	&& mv objs/ngx_http_xslt_filter_module.so objs/ngx_http_xslt_filter_module-debug.so \
 	&& mv objs/ngx_http_image_filter_module.so objs/ngx_http_image_filter_module-debug.so \
+	&& mv objs/ngx_http_geoip_module.so objs/ngx_http_geoip_module-debug.so \
 	&& mv objs/ngx_http_perl_module.so objs/ngx_http_perl_module-debug.so \
 	&& mv objs/ngx_stream_geoip_module.so objs/ngx_stream_geoip_module-debug.so \
 	&& ./configure $CONFIG \
@@ -110,6 +147,7 @@ RUN GPG_KEYS=B0F4253373F8F6F510D42178520A9993A1C052F8 \
 	&& install -m755 objs/nginx-debug /usr/sbin/nginx-debug \
 	&& install -m755 objs/ngx_http_xslt_filter_module-debug.so /usr/lib/nginx/modules/ngx_http_xslt_filter_module-debug.so \
 	&& install -m755 objs/ngx_http_image_filter_module-debug.so /usr/lib/nginx/modules/ngx_http_image_filter_module-debug.so \
+	&& install -m755 objs/ngx_http_geoip_module-debug.so /usr/lib/nginx/modules/ngx_http_geoip_module-debug.so \
 	&& install -m755 objs/ngx_http_perl_module-debug.so /usr/lib/nginx/modules/ngx_http_perl_module-debug.so \
 	&& install -m755 objs/ngx_stream_geoip_module-debug.so /usr/lib/nginx/modules/ngx_stream_geoip_module-debug.so \
 	&& ln -s ../../usr/lib/nginx/modules /etc/nginx/modules \
@@ -135,14 +173,17 @@ RUN GPG_KEYS=B0F4253373F8F6F510D42178520A9993A1C052F8 \
 	&& apk del .build-deps \
 	&& apk del .gettext \
 	&& mv /tmp/envsubst /usr/local/bin/ \
+	&& rm -rf /tmp/luarocks-* /tmp/LuaJIT-* \
 	\
 	# forward request and error logs to docker log collector
 	&& ln -sf /dev/stdout /var/log/nginx/access.log \
-	&& ln -sf /dev/stderr /var/log/nginx/error.log
+	&& ln -sf /dev/stderr /var/log/nginx/error.log 
 
 COPY nginx.conf /etc/nginx/nginx.conf
 COPY nginx.vh.default.conf /etc/nginx/conf.d/default.conf
 
-EXPOSE 80 443
+EXPOSE 80
+
+STOPSIGNAL SIGQUIT
 
 CMD ["nginx", "-g", "daemon off;"]
